+++ test/denormalize-tests/with-partial-false.ts
@@ -28,7 +28,7 @@ export const test: DenormalizeOneTest = {
   `,
   partial: false,
   stale: false,
-  staleEntities: {},
+  staleMap: {},
   data: {
     posts: [
       {
@@ -44,7 +44,7 @@ export const test: DenormalizeOneTest = {
       }
     ]
   },
-  entities: {
+  normMap: {
     ROOT_QUERY: {
       posts: ["Post;123"]
     },
