+++ test/shared-tests/query-with-array-of-string.ts
@@ -21,7 +21,7 @@ export const test: OneTest = {
       }
     ]
   },
-  entities: {
+  normMap: {
     ROOT_QUERY: {
       posts: ["Post;123"]
     },
