+++ test/shared-tests/same-entity-twice-different-fields.ts
@@ -53,7 +53,7 @@ export const test: OneTest = {
       }
     ]
   },
-  entities: {
+  normMap: {
     ROOT_QUERY: {
       posts: ["Post;123"]
     },
