+++ test/shared-tests/with-deep-null-values.ts
@@ -41,7 +41,7 @@ export const test: OneTest = {
       }
     ]
   },
-  entities: {
+  normMap: {
     ROOT_QUERY: {
       posts: ["Post;123"]
     },
