+++ test/shared-tests/with-null-values.ts
@@ -47,7 +47,7 @@ export const test: OneTest = {
       }
     ]
   },
-  entities: {
+  normMap: {
     ROOT_QUERY: {
       posts: ["Post;123"]
     },
