+++ test/shared-tests/with-reserved-words.ts
@@ -48,7 +48,7 @@ export const test: OneTest = {
       }
     ]
   },
-  entities: {
+  normMap: {
     ROOT_QUERY: {
       posts: ["Post;123"]
     },
